Update CI to avoid `tox -e ALL`
When this is used, it means that no new toxenv can be introduced which
is incompatible with pypy. This poses an issue for work using mypy to
check annotations, as pypy currently cannot install it.

In order to move away from `-e ALL`, expand the github actions matrix
config to generate the list of desirable runs. The TOXENV=py setting
is used to run the main testenv on the current interpreter, and the
TOXENV=pyparsing_packaging builds (which run today) are preserved
except for on macos and pypy.

The codecov step now looks for `TOXENV=py` in addition to other matrix
attributes.