ParserElement repr() implementation, with literals
Introduces a _make_repr() method with parameters to account for the
different contexts in which an element may need to be represented (top
level or as a term in an expression, as a potential string literal or
not).  In order to make the code a little easier to read, the
__format__() method is implemented as well, with conversion specifiers
used to indicate the context.