Added a new example showcasing a recursive parser for parsing advanced mathematical expressions in infix notation.
The parser supports binary, unary, and variadic operators that can be customized by modifying the class variables.
It therefore goes way beyond any existing examples for artithmetic expression parsers. It also
showcases how regex can be utilized to improve the performance of a parser written utilzing pyparsing.

I wrote this parser for another project, where it was needed. I could not find a ready-made parser for my purposes in the
examples, or the web in general. Therefore, I am making this pull request in hopes it gets accepted and hopefully lessens
the burden of some other future user, who might have a similar need for a parser as I did. I left comments in the code that I made
for myself when learning the library. I believe they add to the example, and therefore left them there.