Docs: Move __init__ docstrings to class level
Eliminate any docstrings on `__init__` methods, either adding the
contents to the class docstring (`pyparsing.core.Regex`), or making
it the previously-nonexistent class docstring instead
(`pyparsing.core.DelimitedList`).

This allows `__init__` methods to be excluded from the API docs,
as they're redundant with the class-level documentation (which
includes the `__init__` signature).